requirements.txt
phoenix6==25.0.0b3
pynetworktables==2021.0.0
robotpy_commands_v2==2025.0.0b1
robotpy_wpimath==2025.0.0b2
wpilib==2025.0.0b2
limelightlib-python
